Very Important Bug Fix
I did not expect the designer creating custom code for every property I
tried to set manually in the code according to the systems theme. This
was a huge problem because the control kept using the same style
information as on the developer's system. To prevent that, I had to
disable the designer serialization for certain control properties. That
also means that you can't set those properties manually in the designer
but I think you can live without that feature until I have found a
solution to this problem, too.